Rename filed *.ts to *.component.ts and added user models. Changed router place
