+++ application/src/app/footer.component.ts
@@ -2,6 +2,6 @@ import {Component} from '@angular/core';
 
 @Component({
   selector: 'lugbench-footer',
-  template: require('./footer.html')
+  template: require('./footer.component.html')
 })
 export class FooterComponent {}
