--- application/src/app/footer.spec.ts
@@ -1,20 +1,0 @@
-import {FooterComponent} from './footer';
-import {TestBed, async} from '@angular/core/testing';
-
-describe('footer component', () => {
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [
-        FooterComponent
-      ]
-    });
-    TestBed.compileComponents();
-  }));
-
-  it('should render footer', () => {
-    const fixture = TestBed.createComponent(FooterComponent);
-    fixture.detectChanges();
-    const footer = fixture.nativeElement;
-    expect(footer.querySelector('lugbench-footer')).toBeDefined();
-  });
-});
