+++ application/src/app/gpus/gpu.component.html
@@ -1,0 +1,3 @@
+<div *ngIf="gpu">
+  {{ gpu.properties.deviceName }}
+</div>
