+++ application/src/app/gpus/gpus.component.html
