+++ application/src/app/gpus/gpus.component.spec.ts
