+++ application/src/app/main.component.html
@@ -2,7 +2,7 @@
   <lugbench-header></lugbench-header>
   <main class="main">
     <lugbench-title></lugbench-title>
-    <lugbench-gpus></lugbench-gpus>
+    <lugbench-root></lugbench-root>
   </main>
   <lugbench-footer></lugbench-footer>
 </div>
