+++ application/src/app/main.component.ts
@@ -2,6 +2,6 @@ import {Component} from '@angular/core';
 
 @Component({
   selector: 'lugbench-app',
-  template: require('./main.html')
+  template: require('./main.component.html')
 })
 export class MainComponent {}
