+++ application/src/app/shared/models/gpu.ts
@@ -1,0 +1,8 @@
+export class Gpu {
+  public extensions: Array<Object>;
+    public features: Object;
+    public memory: Object;
+    public properties: Object;
+    public formats: Array<Array<Object>>;
+    public queues: Array<Object>;
+}
