+++ application/src/app/title.component.ts
@@ -2,6 +2,6 @@ import {Component} from '@angular/core';
 
 @Component({
   selector: 'lugbench-title',
-  template: require('./title.html')
+  template: require('./title.component.html')
 })
 export class TitleComponent {}
