+++ application/src/index.html
@@ -15,6 +15,6 @@
   </head>
 
   <body>
-    <lugbench-root>Loading...</lugbench-root>
+    <lugbench-app>Loading...</lugbench-app>
   </body>
 </html>
